# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A RESTful API for playing Snake and Ladder game",
        "title": "Snake and Ladder Game API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/game/roll": {
            "post": {
                "description": "Rolls a dice and updates the current player's position",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Roll dice for the current player",
                "parameters": [
                    {
                        "description": "Roll Dice Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RollDiceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GameResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.GameResponse"
                        }
                    }
                }
            }
        },
        "/game/{gameId}": {
            "get": {
                "description": "Retrieves the current state of a game",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Get game status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "gameId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GameResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.GameResponse"
                        }
                    }
                }
            }
        },
        "/lobby/create": {
            "post": {
                "description": "Creates a new game lobby and returns the lobby ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lobby"
                ],
                "summary": "Create a new game lobby",
                "parameters": [
                    {
                        "description": "Create Lobby Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateLobbyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.GameResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.GameResponse"
                        }
                    }
                }
            }
        },
        "/lobby/join": {
            "post": {
                "description": "Allows a player to join an existing lobby using the lobby ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lobby"
                ],
                "summary": "Join an existing game lobby",
                "parameters": [
                    {
                        "description": "Join Lobby Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.JoinLobbyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GameResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.GameResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Board": {
            "type": "object",
            "properties": {
                "ladders": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "size": {
                    "type": "integer",
                    "example": 100
                },
                "snakes": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        },
        "models.CreateLobbyRequest": {
            "type": "object",
            "required": [
                "playerName"
            ],
            "properties": {
                "playerName": {
                    "type": "string",
                    "example": "John"
                }
            }
        },
        "models.Game": {
            "type": "object",
            "properties": {
                "board": {
                    "$ref": "#/definitions/models.Board"
                },
                "currentTurn": {
                    "type": "integer",
                    "example": 0
                },
                "gameStatus": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.GameStatus"
                        }
                    ],
                    "example": "active"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Player"
                    }
                },
                "winner": {
                    "$ref": "#/definitions/models.Player"
                }
            }
        },
        "models.GameResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request format"
                },
                "game": {
                    "$ref": "#/definitions/models.Game"
                },
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                }
            }
        },
        "models.GameStatus": {
            "type": "string",
            "enum": [
                "waiting",
                "active",
                "completed"
            ],
            "x-enum-varnames": [
                "StatusWaiting",
                "StatusActive",
                "StatusCompleted"
            ]
        },
        "models.JoinLobbyRequest": {
            "type": "object",
            "required": [
                "lobbyId",
                "playerName"
            ],
            "properties": {
                "lobbyId": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "playerName": {
                    "type": "string",
                    "example": "Jane"
                }
            }
        },
        "models.Player": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "player-123"
                },
                "name": {
                    "type": "string",
                    "example": "John"
                },
                "position": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.RollDiceRequest": {
            "type": "object",
            "required": [
                "gameId",
                "playerId"
            ],
            "properties": {
                "gameId": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "playerId": {
                    "type": "string",
                    "example": "player-123"
                }
            }
        }
    }
}